pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "pg-advisor"
version     = "0.1.0"
description = "PostgreSQL schema & query advisor — rule-based, no AI"
readme      = "README.md"
license     = { text = "MIT" }
requires-python = ">=3.11"

keywords = [
  "postgresql", "postgres", "database",
  "schema", "advisor", "linter", "index"
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Database",
]

dependencies = [
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov",
  "build",
  "twine",
]

[project.scripts]
pg-advisor = "pg_advisor.cli:main"

[project.urls]
Homepage   = "https://github.com/Salman-Aijaz/pg_advisor"
Issues     = "https://github.com/Salman-Aijaz/pg_advisor"

[tool.setuptools.packages.find]
where = ["."]
include = ["pg_advisor", "pg_advisor.*"]